requirements.txt
openai>=1.0.0
python-docx>=0.8.11
pyyaml>=6.0
tiktoken>=0.5.0
